% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_comparison_boxplot.R
\name{Model_comparison_boxplot}
\alias{Model_comparison_boxplot}
\title{Makes a boxplot comparing the R2 of different Pcf fitted model}
\usage{
Model_comparison_boxplot(
  list_results = list(Results_power_law, Results_exponential)
)
}
\arguments{
\item{list_results}{A list containing the results of different fitted models}
}
\value{
NULL This function does not return a value; it displays a plot.
}
\description{
Makes a boxplot comparing the R2 of different Pcf fitted model
}
